+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/FoundryProperties.kt
@@ -70,7 +70,11 @@ internal constructor(
 
   /** Indicates that this android library project has variants. Flag-only, value is ignored. */
   public val libraryWithVariants: Boolean
-    get() = booleanProperty("slack.gradle.config.libraryWithVariants")
+    get() = booleanProperty("foundry.gradle.config.libraryWithVariants")
+
+  /** Default namespace prefix for android proejcts if one isn't specified. */
+  public val defaultNamespacePrefix: String
+    get() = stringProperty("foundry.android.defaultNamespacePrefix")
 
   /**
    * Indicates that the gradle versions plugin should allow unstable versions. By default unstable
@@ -78,74 +82,74 @@ internal constructor(
    * is ignored.
    */
   public val versionsPluginAllowUnstable: Boolean
-    get() = booleanProperty("slack.gradle.config.versionsPluginAllowUnstable")
+    get() = booleanProperty("foundry.gradle.config.versionsPluginAllowUnstable")
 
   /** Opt-out flag to skip the androidx dependency check. Should only be used for debugging. */
   public val skipAndroidxCheck: Boolean
-    get() = booleanProperty("slack.gradle.skipAndroidXCheck")
+    get() = booleanProperty("foundry.gradle.skipAndroidXCheck")
 
   /** Version code used for debug APK outputs. */
   public val debugVersionCode: Int
-    get() = intProperty("slack.gradle.debugVersionCode", 90009999)
+    get() = intProperty("foundry.gradle.debugVersionCode", 90009999)
 
   /** User string used for debug APK outputs. */
   public val debugUserString: String
-    get() = stringProperty("slack.gradle.debugUserString", "debug")
+    get() = stringProperty("foundry.gradle.debugUserString", "debug")
 
   /** Opt-in flag to enable snapshots repos, used for the dependencies build shadow job. */
   public val enableSnapshots: Boolean
-    get() = booleanProperty("slack.gradle.config.enableSnapshots")
+    get() = booleanProperty("foundry.gradle.config.enableSnapshots")
 
   /** Opt-in flag to enable mavenLocal repos, used for local testing. */
   public val enableMavenLocal: Boolean
-    get() = booleanProperty("slack.gradle.config.enableMavenLocal")
+    get() = booleanProperty("foundry.gradle.config.enableMavenLocal")
 
   /**
    * Flag to indicate that this project should have no api dependencies, such as if it's solely an
    * annotation processor.
    */
   public val rakeNoApi: Boolean
-    get() = booleanProperty("slack.gradle.config.rake.noapi")
+    get() = booleanProperty("foundry.gradle.config.rake.noapi")
 
   /**
    * Flag to enable the Gradle Dependency Analysis Plugin, which is disabled by default due to
    * https://github.com/autonomousapps/dependency-analysis-android-gradle-plugin/issues/204
    */
   public val enableAnalysisPlugin: Boolean
-    get() = booleanProperty("slack.gradle.config.enableAnalysisPlugin")
+    get() = booleanProperty("foundry.gradle.config.enableAnalysisPlugin")
 
   /**
    * Flag to indicate this project should be exempted from platforms, usually platform projects
    * themselves.
    */
   public val noPlatform: Boolean
-    get() = booleanProperty("slack.gradle.config.noPlatform")
+    get() = booleanProperty("foundry.gradle.config.noPlatform")
 
   /** Property corresponding to the supported languages in GA builds */
   public val supportedLanguages: String
-    get() = stringProperty("slack.supportedLanguages")
+    get() = stringProperty("foundry.supportedLanguages")
 
   /** Property corresponding to the supported languages in Internal builds */
   public val supportedLanguagesInternal: String
-    get() = stringProperty("slack.supportedLanguagesInternal")
+    get() = stringProperty("foundry.supportedLanguagesInternal")
 
   /** Property corresponding to the supported languages in Beta builds */
   public val supportedLanguagesBeta: String
-    get() = stringProperty("slack.supportedLanguagesBeta")
+    get() = stringProperty("foundry.supportedLanguagesBeta")
 
   /**
    * Property corresponding to the file path of a custom versions.json file for use with
    * dependencies shadow jobs.
    */
   public val versionsJson: File?
-    get() = fileProperty("slack.versionsJson")
+    get() = fileProperty("foundry.versionsJson")
 
   /**
    * An alias name to a libs.versions.toml bundle for common Android Compose dependencies that
    * should be added to android projects with compose enabled
    */
   public val defaultComposeAndroidBundleAlias: String?
-    get() = optionalStringProperty("slack.compose.android.defaultBundleAlias")
+    get() = optionalStringProperty("foundry.compose.android.defaultBundleAlias")
 
   /**
    * Enables live literals. Note that they are disabled by default due to
@@ -153,7 +157,7 @@ internal constructor(
    * https://issuetracker.google.com/issues/274231394.
    */
   public val composeEnableLiveLiterals: Boolean
-    get() = booleanProperty("slack.compose.android.enableLiveLiterals", false)
+    get() = booleanProperty("foundry.compose.android.enableLiveLiterals", false)
 
   /**
    * Common compose compiler options.
@@ -164,14 +168,14 @@ internal constructor(
   public val composeCommonCompilerOptions: Provider<List<String>>
     get() =
       resolver
-        .providerFor("sgp.compose.commonCompilerOptions")
+        .providerFor("foundry.compose.commonCompilerOptions")
         .map { value -> value.split(",").map { it.trim() } }
         .orElse(emptyList())
 
   /** Relative path to a Compose stability configuration file from the _root_ project. */
   public val composeStabilityConfigurationPath: Provider<RegularFile>
     get() =
-      resolver.providerFor("sgp.compose.stabilityConfigurationPath").map {
+      resolver.providerFor("foundry.compose.stabilityConfigurationPath").map {
         project.rootProject.layout.projectDirectory.file(it)
       }
 
@@ -183,7 +187,7 @@ internal constructor(
    */
   public val composeUseIncludeInformationWorkaround: Boolean
     get() =
-      resolver.booleanValue("sgp.compose.useIncludeInformationWorkaround", defaultValue = true)
+      resolver.booleanValue("foundry.compose.useIncludeInformationWorkaround", defaultValue = true)
 
   /**
    * By default, Compose on android only enables source information in debug variants. This is a bit
@@ -196,7 +200,7 @@ internal constructor(
   public val composeIncludeSourceInformationEverywhereByDefault: Boolean
     get() =
       resolver.booleanValue(
-        "sgp.compose.includeSourceInformationEverywhereByDefault",
+        "foundry.compose.includeSourceInformationEverywhereByDefault",
         defaultValue = true,
       )
 
@@ -210,7 +214,7 @@ internal constructor(
    * source sets), we swap the application id suffix at configuration time.
    */
   public val usePrototypeAppId: Boolean
-    get() = presenceProperty("slack.usePrototypeAppId")
+    get() = presenceProperty("foundry.usePrototypeAppId")
 
   /**
    * Property corresponding to the SDK versions we test in Robolectric tests. Its value should be a
@@ -218,19 +222,19 @@ internal constructor(
    */
   public val robolectricTestSdks: List<Int>
     get() =
-      stringProperty("slack.robolectricTestSdks").splitToSequence(",").map { it.toInt() }.toList()
+      stringProperty("foundry.robolectricTestSdks").splitToSequence(",").map { it.toInt() }.toList()
 
   /** Property corresponding to the preinstrumented jars version (the `-i2` suffix in jars). */
   public val robolectricIVersion: Int
-    get() = intProperty("slack.robolectricIVersion")
+    get() = intProperty("foundry.robolectricIVersion")
 
   /** Opt out for -Werror. */
   public val allowWarnings: Provider<Boolean>
-    get() = resolver.booleanProvider("sgp.kotlin.allowWarnings", defaultValue = false)
+    get() = resolver.booleanProvider("foundry.kotlin.allowWarnings", defaultValue = false)
 
   /** Opt out for -Werror in tests. */
   public val allowWarningsInTests: Provider<Boolean>
-    get() = resolver.booleanProvider("sgp.kotlin.allowWarningsInTests", defaultValue = false)
+    get() = resolver.booleanProvider("foundry.kotlin.allowWarningsInTests", defaultValue = false)
 
   /**
    * Anvil generator projects that should always be included when Anvil is enabled.
@@ -238,7 +242,7 @@ internal constructor(
    * This should be semicolon-delimited Gradle project paths.
    */
   public val anvilGeneratorProjects: String?
-    get() = optionalStringProperty("slack.anvil.generatorProjects")
+    get() = optionalStringProperty("foundry.anvil.generatorProjects")
 
   /**
    * Anvil runtime projects that should always be included when Anvil is enabled.
@@ -246,22 +250,22 @@ internal constructor(
    * This should be semicolon-delimited Gradle project paths.
    */
   public val anvilRuntimeProjects: String?
-    get() = optionalStringProperty("slack.anvil.runtimeProjects")
+    get() = optionalStringProperty("foundry.anvil.runtimeProjects")
 
   /** Flag to enable use of the Anvil KSP fork. https://github.com/ZacSweers/anvil */
   public val anvilUseKspFork: Boolean
-    get() = booleanProperty("sgp.anvil.useKspFork", defaultValue = false)
+    get() = booleanProperty("foundry.anvil.useKspFork", defaultValue = false)
 
   /** Log Slack extension configuration state verbosely. */
-  public val slackExtensionVerbose: Boolean
-    get() = booleanProperty("slack.extension.verbose")
+  public val foundryExtensionVerbose: Boolean
+    get() = booleanProperty("foundry.extension.verbose")
 
   /**
    * Flag for Error-Prone auto-patching. Enable when running an auto-patch of EP, such as when it's
    * being introduced to a new module or upgrading EP itself.
    */
   public val errorProneAutoPatch: Boolean
-    get() = booleanProperty("slack.epAutoPatch")
+    get() = booleanProperty("foundry.epAutoPatch")
 
   /**
    * Error-Prone checks that should be considered errors.
@@ -271,14 +275,14 @@ internal constructor(
    * Example: "AnnotationMirrorToString:AutoValueSubclassLeaked"
    */
   public val errorProneCheckNamesAsErrors: String?
-    get() = optionalStringProperty("slack.epCheckNamesAsErrors")
+    get() = optionalStringProperty("foundry.epCheckNamesAsErrors")
 
   /**
    * Flag for Nullaway baselining. When enabled along with [errorProneAutoPatch], existing
    * nullability issues will be baselined with a `castToNonNull` call to wrap it.
    */
   public val nullawayBaseline: Boolean
-    get() = booleanProperty("slack.nullaway.baseline")
+    get() = booleanProperty("foundry.nullaway.baseline")
 
   /**
    * Ndk version to use for android projects.
@@ -286,18 +290,18 @@ internal constructor(
    * Latest versions can be found at https://developer.android.com/ndk/downloads
    */
   public val ndkVersion: String?
-    get() = optionalStringProperty("slack.ndkVersion")
+    get() = optionalStringProperty("foundry.ndkVersion")
 
   /**
    * Enables verbose logging in miscellaneous places of SGP. This is intended to be a less noisy
    * alternative to running gradle with `--info` or `--debug`.
    */
   public val verboseLogging: Boolean
-    get() = resolver.booleanValue("sgp.logging.verbose")
+    get() = resolver.booleanValue("foundry.logging.verbose")
 
   /** Flag to enable verbose logging in unit tests. */
   public val testVerboseLogging: Boolean
-    get() = booleanProperty("slack.test.verboseLogging") || verboseLogging
+    get() = booleanProperty("foundry.test.verboseLogging") || verboseLogging
 
   /**
    * Flag to enable kapt in tests. By default these are disabled due to this undesirable (but
@@ -307,37 +311,37 @@ internal constructor(
    * See https://youtrack.jetbrains.com/issue/KT-29481#focus=Comments-27-4651462.0-0
    */
   public val enableKaptInTests: Boolean
-    get() = booleanProperty("slack.enabled-kapt-in-tests")
+    get() = booleanProperty("foundry.enabled-kapt-in-tests")
 
   /** Flag to enable errors only in lint checks. */
   public val lintErrorsOnly: Boolean
-    get() = booleanProperty("slack.lint.errors-only")
+    get() = booleanProperty("foundry.lint.errors-only")
 
   /** File name to use for a project's lint baseline. */
   public val lintBaselineFileName: String?
-    get() = optionalStringProperty("slack.lint.baseline-file-name", blankIsNull = true)
+    get() = optionalStringProperty("foundry.lint.baseline-file-name", blankIsNull = true)
 
   /** Flag to control whether or not lint checks test sources. */
   public val lintCheckTestSources: Boolean
-    get() = booleanProperty("sgp.lint.checkTestSources", true)
+    get() = booleanProperty("foundry.lint.checkTestSources", true)
 
   /** Flag to control whether or not lint checks ignores test sources. */
   public val lintIgnoreTestSources: Boolean
-    get() = booleanProperty("sgp.lint.ignoreTestSources", false)
+    get() = booleanProperty("foundry.lint.ignoreTestSources", false)
 
   /**
    * Flag to control which agp version should be used for lint. Optional. Value should be a version
    * key in `libs.versions.toml`,
    */
   public val lintVersionOverride: String?
-    get() = optionalStringProperty("sgp.lint.agpVersion")
+    get() = optionalStringProperty("foundry.lint.agpVersion")
 
   /**
    * Flag to indicate whether this project is a test library (such as test utils, test fixtures,
    * etc).
    */
   public val isTestLibrary: Boolean
-    get() = booleanProperty("sgp.isTestLibrary", false) || project.name == "test-fixtures"
+    get() = booleanProperty("foundry.isTestLibrary", false) || project.name == "test-fixtures"
 
   /**
    * At the time of writing, AGP does not support running lint on `com.android.test` projects. This
@@ -346,15 +350,15 @@ internal constructor(
    * https://issuetracker.google.com/issues/208765813
    */
   public val enableLintInAndroidTestProjects: Boolean
-    get() = booleanProperty("sgp.lint.enableOnAndroidTestProjects", false)
+    get() = booleanProperty("foundry.lint.enableOnAndroidTestProjects", false)
 
   /** Flag to enable/disable KSP. */
   public val allowKsp: Boolean
-    get() = booleanProperty("slack.allow-ksp")
+    get() = booleanProperty("foundry.allow-ksp")
 
   /** Flag to enable/disable Moshi-IR. */
   public val allowMoshiIr: Boolean
-    get() = booleanProperty("slack.allow-moshi-ir")
+    get() = booleanProperty("foundry.allow-moshi-ir")
 
   /** Flag to enable/disable moshi proguard rule gen. */
   public val moshixGenerateProguardRules: Boolean
@@ -362,15 +366,15 @@ internal constructor(
 
   /** Flag to connect SqlDelight sources to KSP. */
   public val kspConnectSqlDelight: Boolean
-    get() = booleanProperty("sgp.ksp.connect.sqldelight")
+    get() = booleanProperty("foundry.ksp.connect.sqldelight")
 
   /** Flag to connect ViewBinding sources to KSP. */
   public val kspConnectViewBinding: Boolean
-    get() = booleanProperty("sgp.ksp.connect.viewbinding")
+    get() = booleanProperty("foundry.ksp.connect.viewbinding")
 
   /** Variants that should be disabled in a given subproject. */
   public val disabledVariants: String?
-    get() = optionalStringProperty("slack.disabledVariants")
+    get() = optionalStringProperty("foundry.disabledVariants")
 
   /**
    * The Slack-specific kotlin.daemon.jvmargs computed by bootstrap.
@@ -388,15 +392,15 @@ internal constructor(
    * the unit test task for a single build variant (e.g. "testReleaseUnitTest").
    */
   public val ciUnitTestEnabled: Boolean
-    get() = booleanProperty("slack.ci-unit-test.enable", defaultValue = true)
+    get() = booleanProperty("foundry.ci-unit-test.enable", defaultValue = true)
 
   /** CI unit test variant (Android only). Defaults to `release`. */
   public val ciUnitTestVariant: String
-    get() = stringProperty("slack.ci-unit-test.variant", "release")
+    get() = stringProperty("foundry.ci-unit-test.variant", "release")
 
   /** If enabled, applies the kotlinx-kover plugin to projects using ciUnitTest. */
   public val ciUnitTestEnableKover: Boolean
-    get() = booleanProperty("slack.ci-unit-test.enableKover", false)
+    get() = booleanProperty("foundry.ci-unit-test.enableKover", false)
 
   /**
    * Parallelism multiplier to use for unit tests. This should be a float value that is multiplied
@@ -404,17 +408,17 @@ internal constructor(
    */
   public val unitTestParallelismMultiplier: Float
     get() {
-      val rawValue = stringProperty("slack.unit-test.parallelismMultiplier", "0.5")
+      val rawValue = stringProperty("foundry.unit-test.parallelismMultiplier", "0.5")
       val floatValue = rawValue.toFloatOrNull()
       require(floatValue != null && floatValue > 0) {
-        "Invalid value for slack.unit-test.parallelismMultiplier: '$rawValue'"
+        "Invalid value for foundry.unit-test.parallelismMultiplier: '$rawValue'"
       }
       return floatValue
     }
 
   /** Controls how often to fork the JVM in unit tests. Default is 1000. */
   public val unitTestForkEvery: Long
-    get() = intProperty("slack.unit-test.forkEvery", 1000).toLong()
+    get() = intProperty("foundry.unit-test.forkEvery", 1000).toLong()
 
   /**
    * Flag to enable ciLint on a project. Default is true.
@@ -423,14 +427,14 @@ internal constructor(
    * all the lint tasks in the project.
    */
   public val ciLintEnabled: Boolean
-    get() = booleanProperty("slack.ci-lint.enable", defaultValue = true)
+    get() = booleanProperty("foundry.ci-lint.enable", defaultValue = true)
 
   /**
    * Comma-separated list of CI lint variants to run (Android only). Default when unspecified will
    * lint all variants.
    */
   public val ciLintVariants: String?
-    get() = optionalStringProperty("slack.ci-lint.variants")
+    get() = optionalStringProperty("foundry.ci-lint.variants")
 
   /** Flag for enabling test orchestrator. */
   public val useOrchestrator: Boolean
@@ -443,17 +447,17 @@ internal constructor(
    * Should be `:path:to:robolectric-core` format
    */
   public val robolectricCoreProject: String?
-    get() = optionalStringProperty("slack.location.robolectric-core")
+    get() = optionalStringProperty("foundry.location.robolectric-core")
 
   /**
    * Gradle path to a platform project to be referenced by other projects.
    *
-   * Should be `:path:to:slack-platform` format
+   * Should be `:path:to:foundry-platform` format
    *
    * @see Platforms
    */
   public val platformProjectPath: String?
-    get() = optionalStringProperty("slack.location.slack-platform")
+    get() = optionalStringProperty("foundry.location.foundry-platform")
 
   /**
    * Opt-in path for commit hooks in the consuming repo that should be automatically installed
@@ -462,7 +466,7 @@ internal constructor(
    * Corresponds to git's `core.hooksPath`.
    */
   public val gitHooksFile: File?
-    get() = fileProperty("slack.git.hooksPath")
+    get() = fileProperty("foundry.git.hooksPath")
 
   /**
    * Opt-in path for a pre-commit hook in the consuming repo that should be automatically installed
@@ -471,39 +475,39 @@ internal constructor(
    * Corresponds to git's `blame.ignoreRevsFile`.
    */
   public val gitIgnoreRevsFile: File?
-    get() = fileProperty("slack.git.ignoreRevsFile")
+    get() = fileProperty("foundry.git.ignoreRevsFile")
 
   /**
    * Optional file location for an `affected_projects.txt` file that contains a list of projects
    * affected in this build.
    */
   public val affectedProjects: File?
-    get() = fileProperty("slack.avoidance.affectedProjectsFile")
+    get() = fileProperty("foundry.avoidance.affectedProjectsFile")
 
   /* Controls for Java/JVM/JDK versions uses in compilations and execution of tests. */
 
   /** Flag to enable strict JDK mode, forcing some things like JAVA_HOME. */
   public val strictJdk: Boolean
-    get() = booleanProperty("slackToolchainsStrict", defaultValue = true)
+    get() = booleanProperty("foundryToolchainsStrict", defaultValue = true)
 
   /** Android cache fix plugin. */
-  public val enableAndroidCacheFix: Boolean = booleanProperty("slack.plugins.android-cache-fix")
+  public val enableAndroidCacheFix: Boolean = booleanProperty("foundry.plugins.android-cache-fix")
 
   /**
    * Optional override for buildToolsVersion in Android projects. Sometimes temporarily necessary to
    * pick up new fixes.
    */
   public val buildToolsVersionOverride: String? =
-    optionalStringProperty("sgp.android.buildToolsVersionOverride")
+    optionalStringProperty("foundry.android.buildToolsVersionOverride")
 
   /**
    * Performance optimization to relocate the entire project build directory to a location outside
    * the IDE's view. This prevents the IDE from tracking these files and improves IDE performance.
    */
-  public val relocateBuildDir: Boolean = betaFeature("sgp.perf.relocateBuildDir")
+  public val relocateBuildDir: Boolean = betaFeature("foundry.perf.relocateBuildDir")
 
   /** Opt-in for beta SGP features. */
-  public val enableBetaFeatures: Boolean = booleanProperty("sgp.beta", defaultValue = false)
+  public val enableBetaFeatures: Boolean = booleanProperty("foundry.beta", defaultValue = false)
 
   /**
    * Shorthand helper for checking features that are in beta or falling back to their specific flag.
@@ -514,22 +518,22 @@ internal constructor(
 
   /* Controls for auto-applied plugins. */
   public val autoApplyTestRetry: Boolean
-    get() = booleanProperty("slack.auto-apply.test-retry", defaultValue = true)
+    get() = booleanProperty("foundry.auto-apply.test-retry", defaultValue = true)
 
   public val autoApplySpotless: Boolean
-    get() = booleanProperty("slack.auto-apply.spotless", defaultValue = true)
+    get() = booleanProperty("foundry.auto-apply.spotless", defaultValue = true)
 
   public val autoApplyDetekt: Boolean
-    get() = booleanProperty("slack.auto-apply.detekt", defaultValue = true)
+    get() = booleanProperty("foundry.auto-apply.detekt", defaultValue = true)
 
   public val autoApplyNullaway: Boolean
-    get() = booleanProperty("slack.auto-apply.nullaway", defaultValue = true)
+    get() = booleanProperty("foundry.auto-apply.nullaway", defaultValue = true)
 
   public val autoApplyCacheFix: Boolean
-    get() = booleanProperty("slack.auto-apply.cache-fix", defaultValue = true)
+    get() = booleanProperty("foundry.auto-apply.cache-fix", defaultValue = true)
 
   public val autoApplySortDependencies: Boolean
-    get() = booleanProperty("slack.auto-apply.sort-dependencies", defaultValue = true)
+    get() = booleanProperty("foundry.auto-apply.sort-dependencies", defaultValue = true)
 
   /* Test retry controls. */
   public enum class TestRetryPluginType {
@@ -539,50 +543,50 @@ internal constructor(
 
   public val testRetryPluginType: TestRetryPluginType
     get() =
-      stringProperty("slack.test.retry.pluginType", TestRetryPluginType.RETRY_PLUGIN.name)
+      stringProperty("foundry.test.retry.pluginType", TestRetryPluginType.RETRY_PLUGIN.name)
         .let(TestRetryPluginType::valueOf)
 
   public val testRetryFailOnPassedAfterRetry: Provider<Boolean>
     get() =
-      resolver.booleanProvider("slack.test.retry.failOnPassedAfterRetry", defaultValue = false)
+      resolver.booleanProvider("foundry.test.retry.failOnPassedAfterRetry", defaultValue = false)
 
   public val testRetryMaxFailures: Provider<Int>
-    get() = resolver.intProvider("slack.test.retry.maxFailures", defaultValue = 20)
+    get() = resolver.intProvider("foundry.test.retry.maxFailures", defaultValue = 20)
 
   public val testRetryMaxRetries: Provider<Int>
-    get() = resolver.intProvider("slack.test.retry.maxRetries", defaultValue = 1)
+    get() = resolver.intProvider("foundry.test.retry.maxRetries", defaultValue = 1)
 
   /* Detekt configs. */
   /** Detekt config files, evaluated from rootProject.file(...). */
   public val detektConfigs: List<String>?
-    get() = optionalStringProperty("slack.detekt.configs")?.split(",")
+    get() = optionalStringProperty("foundry.detekt.configs")?.split(",")
 
   /** Detekt baseline file, evaluated from project.layout.projectDirectory.file(...). */
   public val detektBaselineFileName: String?
-    get() = optionalStringProperty("slack.detekt.baseline-file-name", blankIsNull = true)
+    get() = optionalStringProperty("foundry.detekt.baseline-file-name", blankIsNull = true)
 
   /** Enables full detekt mode (with type resolution). Off by default due to performance issues. */
   public val enableFullDetekt: Boolean
-    get() = booleanProperty("slack.detekt.full")
+    get() = booleanProperty("foundry.detekt.full")
 
   /** Comma-separated set of projects to ignore in sorting dependencies. */
   public val sortDependenciesIgnore: String?
-    get() = optionalStringProperty("slack.sortDependencies.ignore")
+    get() = optionalStringProperty("foundry.sortDependencies.ignore")
 
   /** Enables verbose debug logging across the plugin. */
   public val debug: Boolean
-    get() = booleanProperty("slack.debug", defaultValue = false)
+    get() = booleanProperty("foundry.debug", defaultValue = false)
 
   /** A comma-separated list of configurations to use in affected project detection. */
   public val affectedProjectConfigurations: String?
-    get() = optionalStringProperty("slack.avoidance.affected-project-configurations")
+    get() = optionalStringProperty("foundry.avoidance.affected-project-configurations")
 
   /**
    * Flag to, when true, makes [affectedProjectConfigurations] build upon the defaults rather than
    * replace them.
    */
   public val buildUponDefaultAffectedProjectConfigurations: Boolean
-    get() = booleanProperty("slack.avoidance.build-upon-default-affected-project-configurations")
+    get() = booleanProperty("foundry.avoidance.build-upon-default-affected-project-configurations")
 
   /**
    * Global control for enabling stricter validation of projects, such as ensuring Kotlin projects
@@ -593,30 +597,31 @@ internal constructor(
    * Granular controls should depend on this check + include their own opt-out check as-needed.
    */
   public val strictMode: Boolean
-    get() = booleanProperty("slack.strict", defaultValue = false)
+    get() = booleanProperty("foundry.strict", defaultValue = false)
 
   /** Specific toggle for validating manifests in androidTest sources. */
   public val strictValidateAndroidTestManifest: Boolean
-    get() = booleanProperty("slack.strict.validateAndroidTestManifests", defaultValue = true)
+    get() = booleanProperty("foundry.strict.validateAndroidTestManifests", defaultValue = true)
 
   /**
    * Always enables resources in android unit tests. Only present for benchmarking purposes and
    * should otherwise be off.
    */
   public val alwaysEnableResourcesInTests: Boolean
-    get() = booleanProperty("slack.gradle.config.test.alwaysEnableResources", defaultValue = false)
+    get() =
+      booleanProperty("foundry.gradle.config.test.alwaysEnableResources", defaultValue = false)
 
   /** Global toggle to enable bugsnag. Note this still respects variant filters. */
   public val bugsnagEnabled: Provider<Boolean>
-    get() = resolver.booleanProvider("slack.gradle.config.bugsnag.enabled")
+    get() = resolver.booleanProvider("foundry.gradle.config.bugsnag.enabled")
 
   /** Branch pattern for git branches Bugsnag should be enabled on. */
   public val bugsnagEnabledBranchPattern: Provider<String>
-    get() = resolver.optionalStringProvider("slack.gradle.config.bugsnag.enabledBranchPattern")
+    get() = resolver.optionalStringProvider("foundry.gradle.config.bugsnag.enabledBranchPattern")
 
   /** Global boolean that controls whether mod score is enabled on this project. */
   public val modScoreGlobalEnabled: Boolean
-    get() = resolver.booleanValue("slack.gradle.config.modscore.enabled")
+    get() = resolver.booleanValue("foundry.gradle.config.modscore.enabled")
 
   /**
    * Per-project boolean that allows for excluding this project from mod score.
@@ -624,11 +629,11 @@ internal constructor(
    * Note this should only be applied to projects that cannot be depended on.
    */
   public val modScoreIgnore: Boolean
-    get() = resolver.booleanValue("slack.gradle.config.modscore.ignore")
+    get() = resolver.booleanValue("foundry.gradle.config.modscore.ignore")
 
   /** Experimental flag to enable logging thermal throttling on macOS devices. */
   public val logThermals: Boolean
-    get() = resolver.booleanValue("slack.log-thermals", defaultValue = false)
+    get() = resolver.booleanValue("foundry.log-thermals", defaultValue = false)
 
   /**
    * Enables eager configuration of [SgpArtifact] publishing in subprojects. This is behind a flag
@@ -637,14 +642,14 @@ internal constructor(
    * @see StandardProjectConfigurations.setUpSubprojectArtifactPublishing
    */
   public val eagerlyConfigureArtifactPublishing: Boolean
-    get() = resolver.booleanValue("sgp.artifacts.configure-eagerly", defaultValue = false)
+    get() = resolver.booleanValue("foundry.artifacts.configure-eagerly", defaultValue = false)
 
   /**
    * Force-disables Anvil regardless of `SlackExtension.dagger()` settings, useful for K2 testing
    * where Anvil is unsupported.
    */
   public val disableAnvilForK2Testing: Boolean
-    get() = resolver.booleanValue("sgp.anvil.forceDisable", defaultValue = false)
+    get() = resolver.booleanValue("foundry.anvil.forceDisable", defaultValue = false)
 
   /**
    * Defines the [AnvilMode] to use with this compilation. See the docs on that class for more
@@ -652,21 +657,21 @@ internal constructor(
    */
   public val anvilMode: AnvilMode
     get() =
-      resolver.stringValue("sgp.anvil.mode", defaultValue = AnvilMode.K1_EMBEDDED.name).let {
+      resolver.stringValue("foundry.anvil.mode", defaultValue = AnvilMode.K1_EMBEDDED.name).let {
         AnvilMode.valueOf(it.uppercase(Locale.US))
       }
 
   /** Overrides the kotlin language version if present. */
   public val kaptLanguageVersion: Provider<KotlinVersion>
     get() =
-      resolver.optionalStringProvider("sgp.kapt.languageVersion").map {
+      resolver.optionalStringProvider("foundry.kapt.languageVersion").map {
         KotlinVersion.fromVersion(it)
       }
 
   /** Defines a required vendor for JDK toolchains. */
   public val jvmVendor: Provider<String>
     get() =
-      resolver.optionalStringProvider("sgp.config.jvmVendor").map {
+      resolver.optionalStringProvider("foundry.config.jvmVendor").map {
         if (jvmVendorOptOut) {
           sneakyNull()
         } else {
@@ -676,14 +681,14 @@ internal constructor(
 
   /** Flag to disable JVM vendor setting locally. */
   public val jvmVendorOptOut: Boolean
-    get() = booleanProperty("sgp.config.jvmVendor.optOut", defaultValue = false)
+    get() = booleanProperty("foundry.config.jvmVendor.optOut", defaultValue = false)
 
   /**
    * Option to force a specific kotlin language version. By default defers to the KGP default the
    * build is running with.
    */
   public val kotlinLanguageVersionOverride: Provider<String>
-    get() = resolver.optionalStringProvider("sgp.kotlin.languageVersionOverride")
+    get() = resolver.optionalStringProvider("foundry.kotlin.languageVersionOverride")
 
   /**
    * Free compiler arguments to pass to Kotlin's `freeCompilerArgs` property in all compilations.
@@ -692,7 +697,7 @@ internal constructor(
   public val kotlinFreeArgs: Provider<List<String>>
     get() =
       resolver
-        .optionalStringProvider("sgp.kotlin.freeArgs")
+        .optionalStringProvider("foundry.kotlin.freeArgs")
         .map { it.split(',') }
         // Super important to default if absent due to
         // https://docs.gradle.org/8.7/release-notes.html#build-authoring-improvements
@@ -705,7 +710,7 @@ internal constructor(
   public val kotlinJvmFreeArgs: Provider<List<String>>
     get() =
       resolver
-        .optionalStringProvider("sgp.kotlin.jvmFreeArgs")
+        .optionalStringProvider("foundry.kotlin.jvmFreeArgs")
         .map { it.split(',') }
         // Super important to default if absent due to
         // https://docs.gradle.org/8.7/release-notes.html#build-authoring-improvements
@@ -715,7 +720,7 @@ internal constructor(
   public val kotlinOptIn: Provider<List<String>>
     get() =
       resolver
-        .optionalStringProvider("sgp.kotlin.optIns")
+        .optionalStringProvider("foundry.kotlin.optIns")
         .map { it.split(',') }
         // Super important to default if absent due to
         // https://docs.gradle.org/8.7/release-notes.html#build-authoring-improvements
@@ -723,12 +728,12 @@ internal constructor(
 
   /** Default for Kotlin's `progressive` mode. Defaults to enabled. */
   public val kotlinProgressive: Provider<Boolean>
-    get() = resolver.booleanProvider("sgp.kotlin.progressive", defaultValue = true)
+    get() = resolver.booleanProvider("foundry.kotlin.progressive", defaultValue = true)
 
   internal fun requireAndroidSdkProperties(): AndroidSdkProperties {
-    val compileSdk = compileSdkVersion ?: error("slack.compileSdkVersion not set")
-    val minSdk = minSdkVersion?.toInt() ?: error("slack.minSdkVersion not set")
-    val targetSdk = targetSdkVersion?.toInt() ?: error("slack.targetSdkVersion not set")
+    val compileSdk = compileSdkVersion ?: error("foundry.compileSdkVersion not set")
+    val minSdk = minSdkVersion?.toInt() ?: error("foundry.minSdkVersion not set")
+    val targetSdk = targetSdkVersion?.toInt() ?: error("foundry.targetSdkVersion not set")
     return AndroidSdkProperties(compileSdk, minSdk, targetSdk)
   }
 
@@ -739,16 +744,16 @@ internal constructor(
   )
 
   public val compileSdkVersion: String?
-    get() = optionalStringProperty("slack.compileSdkVersion")
+    get() = optionalStringProperty("foundry.compileSdkVersion")
 
   public fun latestCompileSdkWithSources(defaultValue: Int): Int =
-    intProperty("slack.latestCompileSdkWithSources", defaultValue = defaultValue)
+    intProperty("foundry.latestCompileSdkWithSources", defaultValue = defaultValue)
 
   private val minSdkVersion: String?
-    get() = optionalStringProperty("slack.minSdkVersion")
+    get() = optionalStringProperty("foundry.minSdkVersion")
 
   private val targetSdkVersion: String?
-    get() = optionalStringProperty("slack.targetSdkVersion")
+    get() = optionalStringProperty("foundry.targetSdkVersion")
 
   public companion object {
     /**
@@ -757,17 +762,17 @@ internal constructor(
      * We don't just blanket use `kotlin.daemon.jvmargs` alone because we don't want to pollute
      * other projects.
      */
-    public const val KOTLIN_DAEMON_ARGS_KEY: String = "slack.kotlin.daemon.jvmargs"
+    public const val KOTLIN_DAEMON_ARGS_KEY: String = "foundry.kotlin.daemon.jvmargs"
 
     /** Minimum xmx value for the Gradle daemon. Value is an integer and unit is gigabytes. */
     // Key-only because it's used in a task init without a project instance
-    public const val MIN_GRADLE_XMX: String = "slack.bootstrap.minGradleXmx"
+    public const val MIN_GRADLE_XMX: String = "foundry.bootstrap.minGradleXmx"
 
     /** Minimum xms value for the Gradle daemon. Value is an integer and unit is gigabytes. */
     // Key-only because it's used in a task init without a project instance
-    public const val MIN_GRADLE_XMS: String = "slack.bootstrap.minGradleXms"
+    public const val MIN_GRADLE_XMS: String = "foundry.bootstrap.minGradleXms"
 
-    private const val CACHED_PROVIDER_EXT_NAME = "slack.properties.provider"
+    private const val CACHED_PROVIDER_EXT_NAME = "foundry.properties.provider"
 
     public operator fun invoke(
       project: Project,